+++ docs/changelog.rst
@@ -23,6 +23,8 @@ In progress (unreleased)
 
 * :gh:issue:`1451` :mod:`mitogen`: Refactor module whitelist & blacklist with
   module overrides and blocks. Improve error messages for denied modules.
+* :gh:issue:`1394` :mod:`ansible_mitogen`, :mod:`mitogen`: Add Incus connection
+  support, allowing users who migrated from LXD to Incus to use ansible-mitogen
 
 
 v0.3.41 (2026-02-10)